PWX-31068: Move functions around to fix import cycle issues.
