+++ drivers/storage/portworx/component/auth.go
@@ -183,7 +183,7 @@ func (a *auth) getPrivateKeyOrGenerate(cluster *corev1.StorageCluster, envVarKey
 	// check for pre-configured shared secret
 	for _, envVar := range cluster.Spec.Env {
 		if envVar.Name == envVarKey {
-			privateKey, err = pxutil.GetValueFromEnvVar(context.TODO(), a.k8sClient, &envVar, cluster.Namespace)
+			privateKey, err = k8sutil.GetValueFromEnvVar(context.TODO(), a.k8sClient, &envVar, cluster.Namespace)
 			if err != nil {
 				return "", err
